lab4/judge_system_backend/requirements.txt
asgiref==3.*
Django==3.*
gunicorn==19.*
dj-database-url==0.*
psycopg2-binary==2.8.*
django-cors-headers==3.*
djangorestframework==3.*
djangorestframework-simplejwt==4.*
PyJWT==2.*
pytz==2021.*
sqlparse==0.*
numpy==1.*
pandas==1.*
munkres==1.*
requests==2.*
WeasyPrint==52.*
Markdown==3.*
